updated readme and refactored for container package
